docs(tabset): make the doc of `select()` clearer
The name of the argument, `tabIdx`, lead the reader to think that what must be passed is the index of the tab, instead of its unique id. Renaming to `tabId` and indicating that the id of the tab is what is expected by the method makes it clearer.

The same confusion is also fixed in the tests.

Closes #588