feat(model): introduce 'container' option
This is major refactoring of the NgbModalService that makes it
easier to use (we no longer need to add ngbModalContainer) and
fixes 2 existing issues.

BREAKING CHANGE:

The `ngbModalContainer` directive is no longer needed and was
removed from this project. Just remove any references to the
`<template ngbModalContainer></template>` from your projects.

Closes #1018
Closes #1264

Closes #1373